+++ src/main/java/hu/elte/wr14yr/musicportal/repository/SongCounterRepository.java
@@ -18,5 +18,5 @@ public interface SongCounterRepository extends CrudRepository<SongCounter, Long>
     void deleteAllBySong(Song song);
 
     @Query(value = "UPDATE SONG_COUNTER SET userId = null WHERE USER_ID = :USER_ID", nativeQuery = true)
-    void deleteUserId(@Param("USER_ID") int userId);
+    void deleteUserId(@Param("USER_ID") long userId);
 }
